updating chat memory
