--- examples/chat-memory/nodemon.json
@@ -1,5 +1,0 @@
-{
-  "watch": [".", ".env", "../../packages/**/dist/**"],
-  "exec": "NODE_OPTIONS='--enable-source-maps' tsx --inspect=0.0.0.0:9229 ./src/index.tsx",
-  "ext": "ts, tsx, js, json"
-}
