+++ examples/chat-memory/package.json
@@ -8,23 +8,26 @@
   },
   "description": "Use GenSX to create a chat completion",
   "scripts": {
-    "dev": "tsx ./src/index.tsx",
-    "start": "npx gensx start ./src/workflows.tsx",
-    "build": "tsc",
-    "deploy": "npx gensx deploy -e OPENAI_API_KEY ./src/workflows.tsx"
+    "dev": "pnpm build && node dist/index.js",
+    "start": "npx gensx start ./src/workflows.ts",
+    "build": "tsc -p tsconfig.build.json",
+    "deploy": "npx gensx deploy -e OPENAI_API_KEY ./src/workflows.ts",
+    "prepare": "ts-patch install -s"
   },
   "dependencies": {
-    "@gensx/openai": "^0.1.27",
-    "@gensx/core": "^0.3.12",
-    "@gensx/storage": "^0.0.9",
+    "@ai-sdk/openai": "^1.3.22",
+    "@gensx/core": "alpha",
+    "@gensx/storage": "0.0.12",
+    "@gensx/vercel-ai": "alpha",
+    "ai": "^4.2.10",
     "openai": "^4.87.4",
     "zod": "^3.24.2"
   },
   "devDependencies": {
     "@types/node": "^22",
-    "eslint": "^9.23.0",
-    "nodemon": "^3.1.9",
-    "tsx": "^4.19.2",
+    "ts-function-decorator": "^0.0.1",
+    "ts-function-decorator-ls": "^0.0.1",
+    "ts-patch": "^3.3.0",
     "typescript": "^5.7.2"
   }
 }
