+++ examples/chat-memory/tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "incremental": false,
+    "plugins": [
+      {
+        "transform": "ts-function-decorator",
+        "transformProgram": true
+      }
+    ]
+  }
+}
